Dependencies folder contains the libraries which are needed to make this project work correctly. The TFT screen used for this project was purchased from this link: https://www.dx.com/p/open-smart-touch-screen-expansion-shield-w-touch-pen-for-arduino-2080038.html
It's an OPENSMART 3.2" TFT touch screen with an ILI9327 driver chip on. It's a shield which is just attached to the Arduino Uno. The Game.ino file just uses the custom font found in Adafruit_GFX/Fonts/Anims1.h to draw a simple wall around some bushes with a man character in the top left.